use a strand for all operations
& just support passing an io_context for now. Move most the impl to a .cpp since it's not templated any longer
